cvllm/requirements.txt
pdfplumber==0.11.0
python-docx==1.1.2
pydantic==2.8.2
jinja2==3.1.4
ollama==0.3.0
langchain==0.2.14
langchain-community==0.2.12
fastapi==0.114.0
uvicorn==0.30.6
rich==13.8.0
python-multipart==0.0.9
aiofiles==24.1.0
llama-index-core==0.11.14
llama-index-embeddings-huggingface==0.2.5
sentence-transformers==3.0.1
